Binary improvements

+++ src/bin/replace.rs
@@ -1,13 +1,22 @@
+#[path = "../../common/common.rs"]
+mod common;
 use std::fs::File;
 use std::io::BufReader;
 use std::path::PathBuf;
+use std::process::ExitCode;
 
 use backhand::{FilesystemReader, FilesystemWriter};
 use clap::Parser;
+use common::after_help;
 
 /// tool to replace files in squashfs filesystems
 #[derive(Parser, Debug)]
-#[command(author, version, name = "replace-backhand")]
+#[command(author,
+          version,
+          name = "replace-backhand",
+          after_help = after_help(),
+          max_term_width = 98,
+)]
 struct Args {
     /// Squashfs input image
     image: PathBuf,
@@ -24,7 +33,7 @@ struct Args {
     out: PathBuf,
 }
 
-fn main() {
+fn main() -> ExitCode {
     tracing_subscriber::fmt::init();
 
     let args = Args::parse();
@@ -36,10 +45,15 @@ fn main() {
 
     // Modify file
     let new_file = File::open(&args.file).unwrap();
-    filesystem.replace_file(args.file_path, new_file).unwrap();
+    if let Err(e) = filesystem.replace_file(args.file_path, new_file) {
+        println!("[!] {e}");
+        return ExitCode::FAILURE;
+    }
 
     // write new file
     let mut output = File::create(&args.out).unwrap();
     filesystem.write(&mut output).unwrap();
     println!("replaced file and wrote to {}", args.out.display());
+
+    ExitCode::SUCCESS
 }
